Moving ArrayPayloadOf to test source
